requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# LLM Providers
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0

# Web Scraping
playwright>=1.41.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
readability-lxml>=0.8.1
httpx>=0.26.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
aiosqlite>=0.19.0

# Redis Queue
redis>=5.0.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# MCP (Model Context Protocol)
mcp>=1.0.0

# CLI
typer>=0.9.0
rich>=13.7.0
questionary>=2.0.1

# Utilities
tenacity>=8.2.0  # Retry logic
structlog>=24.1.0  # Logging
aiofiles>=23.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
ruff>=0.2.0
mypy>=1.8.0
